Drop scrutinizer input normalization
Change-type: patch
